+++ test/mri/excludes/TestGemRequire.rb
@@ -1,6 +1,4 @@
 exclude :test_activate_via_require_respects_loaded_files, "needs investigation"
 exclude :test_require_when_gem_defined, "needs investigation"
 exclude :test_require_works_after_cleanup, "needs investigation"
-exclude :test_default_gem_and_normal_gem, "needs investigation"
-exclude :test_default_gem_only, "needs investigation"
 exclude :test_no_kernel_require_in_warn_with_uplevel, "needs investigation"